Some changes in authorization logic

+++ routers/groups.py
@@ -19,6 +19,12 @@ async def fetch_groups(db: Session=Depends(get_db), user_object: str = Depends(a
 
 @router.post("/groups", status_code=status.HTTP_201_CREATED, response_model=GetGroupSchema)
 async def create_groups(group_data: CreateGroupSchema, db: Session=Depends(get_db), user_object: str = Depends(auth.get_current_user)):
+    if group_data.owner_id != user_object.id:
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail={"message": "Not authenticated"}
+        )
+    
     group_object = Group(**group_data.model_dump())
     db.add(group_object)
     db.commit()
@@ -31,20 +37,32 @@ async def get_group(group_id: UUID, db: Session=Depends(get_db), user_object: st
     group_object = db.query(Group).filter(Group.id == group_id).first()
 
     if group_object is None:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail={"message": "Group not found!"})
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND, 
+            detail={"message": "Group not found!"}
+        )
     
     return group_object
 
 
 @router.patch("/groups/{group_id}", status_code=status.HTTP_200_OK, response_model=GetGroupSchema)
 async def update_group(group_id: UUID, group_data: UpdateGroupSchema, db: Session=Depends(get_db), user_object: str = Depends(auth.get_current_user)):
+    if group_data.owner_id != user_object.id:
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail={"message": "Not authenticated"}
+        )
+    
     update_group_data = group_data.model_dump(exclude_none=True)
 
     group_query = db.query(Group).filter(Group.id == group_id)
     group_object = group_query.first()
 
     if group_object is None:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail={"message": "Group not found!"})
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND, 
+            detail={"message": "Group not found!"}
+        )
     
     group_query.update(update_group_data)
     db.commit()
@@ -56,8 +74,17 @@ async def update_group(group_id: UUID, group_data: UpdateGroupSchema, db: Sessio
 async def delete_group(group_id: str, db: Session=Depends(get_db), user_object: str = Depends(auth.get_current_user)):
     group_object = db.query(Group).filter(Group.id == group_id).first()
 
+    if group_object.owner_id != user_object.id:
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail={"message": "Not authenticated"}
+        )
+
     if group_object is None:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail={"message": "Group not found!"})
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND, 
+            detail={"message": "Group not found!"}
+        )
     
     db.delete(group_object)
     db.commit()
